django-ope/ope/requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-barcode==0.13.1
pytz==2020.1
reportlab==3.5.48
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
